% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umccrise.R
\name{bcftools_stats_plot}
\alias{bcftools_stats_plot}
\title{Parse bcftools stats File}
\usage{
bcftools_stats_plot(x = NULL)
}
\arguments{
\item{x}{Path to bcftools stats \code{bcftools_stats.txt} file.}
}
\value{
A ggplot2 object.
}
\description{
Parses bcftools stats \code{bcftools_stats.txt} file.
}
