% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord.R
\name{chord_get_genome_obj}
\alias{chord_get_genome_obj}
\title{Get BSgenome Object for CHORD}
\usage{
chord_get_genome_obj(genome = "hg38")
}
\arguments{
\item{genome}{Human genome assembly: hg38 (default), hg19 or GRCh37.}
}
\value{
BSgenome object.
}
\description{
Gets BSgenome object for use in CHORD.
}
\examples{
\dontrun{
chord_get_genome_obj("hg38")
}
}
