% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord.R
\name{chord_mantavcf2df}
\alias{chord_mantavcf2df}
\title{Convert Manta VCF to data.frame}
\usage{
chord_mantavcf2df(in_vcf)
}
\arguments{
\item{in_vcf}{Manta VCF.}
}
\value{
Tibble with two columns: sv_type and sv_len (INFO/SVTYPE and INFO/SVLEN from VCF).
}
\description{
Converts a Manta VCF to a data.frame for processing with CHORD.
}
\examples{
in_vcf <- system.file("extdata/umccrise/sv/manta.vcf.gz", package = "gpgr")
d <- chord_mantavcf2df(in_vcf)
}
