% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\docType{data}
\name{cosmic_signatures_2015}
\alias{cosmic_signatures_2015}
\title{Get COSMIC 2015 Signatures}
\format{
A matrix with 96 rows (one for each somatic mutation type within a specific context)
and 30 columns (one for each signature).
}
\usage{
data(cosmic_signatures_2015)
}
\description{
Get COSMIC 2015 Signatures
}
\keyword{datasets}
