% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\name{count_pieces}
\alias{count_pieces}
\title{Count Number of Parts in a String}
\usage{
count_pieces(x, sep)
}
\arguments{
\item{x}{Input string.}

\item{sep}{Pattern to count for.}
}
\value{
Number of parts.
}
\description{
Counts number of pieces of a string separated by a pattern.
If it's an empty string, returns 0. If the pattern isn't found, returns 1.
If the pattern is found once, returns 2 (two pieces), etc.
}
\examples{
(a <- gpgr:::count_pieces("foo,bar,baz", sep = ","))
(b <- gpgr:::count_pieces("foo", sep = ","))
(k <- gpgr:::count_pieces("", sep = ","))
(m <- gpgr:::count_pieces(",", sep = ","))
}
