% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umccrise.R
\name{dragen_hrd}
\alias{dragen_hrd}
\title{Parse DRAGEN HRD File}
\usage{
dragen_hrd(x = NULL)
}
\arguments{
\item{x}{Path to DRAGEN \code{hrdscore.csv} file.}
}
\value{
Tibble with a single row and the following score columns:
\itemize{
\item HRD (homologous recombination deficiency)
\item LOH (loss of heterozygosity)
\item TAI (telomeric allelic imbalance)
\item LST (large-scale state transitions)
}

If no input is provided, the fields are NA.
}
\description{
Parses DRAGEN \code{hrdscore.csv} file.
}
