% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\docType{data}
\name{EFFECT_ABBREVIATIONS}
\alias{EFFECT_ABBREVIATIONS}
\title{Abbreviations for SV effects}
\format{
Named character vector where names are the original names, and values
are the abbreviated names.
}
\usage{
data(EFFECT_ABBREVIATIONS)
}
\description{
Abbreviations for SV effects
}
\keyword{datasets}
