% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord.R
\name{get_genome_obj}
\alias{get_genome_obj}
\title{Get BSgenome Object}
\usage{
get_genome_obj(genome = "hg38")
}
\arguments{
\item{genome}{Human genome assembly string: hg38 (default), hg19 or GRCh37.}
}
\value{
BSgenome object.
}
\description{
Gets BSgenome object given a human genome assembly string.
}
\examples{
\dontrun{
get_genome_obj("hg38")
}
}
