% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncokb.R
\name{get_oncokb_genes}
\alias{get_oncokb_genes}
\title{Get OncoKB Genes From Somewhere}
\usage{
get_oncokb_genes(x, oncokb_genes)
}
\arguments{
\item{x}{Path to something.}

\item{oncokb_genes}{A tibble of something.}
}
\value{
A vector I think.
}
\description{
Get OncoKB Genes From Somewhere
}
