% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shortcuts.R
\docType{package}
\name{gpgr-package}
\alias{gpgr}
\alias{gpgr-package}
\title{gpgr: Genomics Platform Group Reporting}
\description{
Reports and functions used in the Genomics Platform Group at the University of Melbourne Centre for Cancer Research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://umccr.github.io/gpgr/}
  \item \url{https://github.com/umccr/gpgr}
  \item Report bugs at \url{https://github.com/umccr/gpgr/issues}
}

}
\author{
\strong{Maintainer}: Peter Diakumis \email{peterdiakumis@gmail.com} (\href{https://orcid.org/0000-0002-7502-7545}{ORCID})

}
\keyword{internal}
