% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrdetect.R
\name{hrdetect_prep_cnv}
\alias{hrdetect_prep_cnv}
\title{Prepare PURPLE Somatic CNVs for HRDetect}
\usage{
hrdetect_prep_cnv(x, nm = NULL)
}
\arguments{
\item{x}{Path to \code{purple.cnv.somatic.tsv} file.}

\item{nm}{Sample name.}
}
\value{
Tibble with sample name and HRD-LOH index.
}
\description{
Prepares PURPLE somatic CNVs for use with HRDetect.
}
\examples{
x <- system.file("extdata/purple/purple.cnv.somatic.tsv", package = "gpgr")
(l <- hrdetect_prep_cnv(x, nm = "SampleA"))
}
