% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrdetect.R
\name{hrdetect_prep_snvindel}
\alias{hrdetect_prep_snvindel}
\title{Prepare VCF with SNVs/INDELs for use with HRDetect}
\usage{
hrdetect_prep_snvindel(
  x,
  nm = NULL,
  genome = "hg38",
  outdir = NULL,
  sigsToUse = c(1, 2, 3, 5, 6, 8, 13, 17, 18, 20, 26, 30)
)
}
\arguments{
\item{x}{Path to VCF with SNVs and INDELs.}

\item{nm}{Sample name.}

\item{genome}{Human genome version (default: hg38. hg19 means GRCh37).}

\item{outdir}{Directory to output analysis results.}

\item{sigsToUse}{COSMIC signatures to use.}
}
\value{
List with two elements:
\itemize{
\item snv_results: tibble with exposure score and p-value for chosen signatures.
\item indel_results: tibble with a summary of the count of indels and their proportion.
}
}
\description{
Prepares VCF with SNVs/INDELs for use with HRDetect.
}
\examples{
x <- system.file("extdata/umccrise/snv/somatic-ensemble-PASS.vcf.gz", package = "gpgr")
(l <- hrdetect_prep_snvindel(x, nm = "sampleA", outdir = tempdir()))
}
