% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrdetect.R
\name{hrdetect_prep_sv}
\alias{hrdetect_prep_sv}
\title{Prepare VCF with SVs for use with HRDetect}
\usage{
hrdetect_prep_sv(x, nm = NULL, genome = "hg38")
}
\arguments{
\item{x}{Path to VCF with SVs.}

\item{nm}{Sample name.}

\item{genome}{Human genome version (default: hg38).}
}
\value{
Tibble with counts for each SV category.
}
\description{
Prepare VCF with SVs for use with HRDetect
}
\examples{
x <- system.file("extdata/umccrise/v0.18/sv/manta.vcf.gz", package = "gpgr")
(l <- hrdetect_prep_sv(x, nm = "SampleA"))

}
