% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrdetect.R
\name{hrdetect_read_purple_cnv}
\alias{hrdetect_read_purple_cnv}
\title{Prepare PURPLE Somatic CNVs for HRDetect}
\usage{
hrdetect_read_purple_cnv(x, v = "2.39")
}
\arguments{
\item{x}{Path to \code{purple.cnv.somatic.tsv} file.}

\item{v}{PURPLE version (default: 2.39). Used to determine the column names.}
}
\value{
Tibble containing following columns:
\itemize{
\item Chromosome, chromStart, chromEnd
\item total.copy.number.inTumour
\item minor.copy.number.inTumour
}
}
\description{
Prepares PURPLE somatic CNVs for HRDetect.
}
\examples{
x <- system.file("extdata/purple/v2.39/purple.cnv.somatic.tsv", package = "gpgr")
(cnv <- hrdetect_read_purple_cnv(x))

}
