% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrdetect.R
\name{hrdetect_read_sv_vcf}
\alias{hrdetect_read_sv_vcf}
\title{Read VCF with SVs for use with HRDetect}
\usage{
hrdetect_read_sv_vcf(x, nm = NULL, hg_version = "hg38")
}
\arguments{
\item{x}{Path to VCF.}

\item{nm}{Sample name.}

\item{hg_version}{Human genome version (default: hg38).}
}
\value{
Tibble with following BEDPE-like columns:
\itemize{
\item chrom1
\item start1
\item end1
\item chrom2
\item start2
\item end2
\item sample
\item strand1
\item strand2
}
}
\description{
Reads a VCF with SVs for use with HRDetect.
}
\examples{
x <- system.file("extdata/umccrise/v0.18/sv/manta.vcf.gz", package = "gpgr")
sv_bedpe <- hrdetect_read_sv_vcf(x, nm = "SAMPLE")
head(sv_bedpe)

}
