% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_vcf}
\alias{is_vcf}
\title{Does the input file look like a VCF?}
\usage{
is_vcf(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
TRUE if the input file is inferred to be a VCF, FALSE otherwise.
}
\description{
Quickly checks that the input file has a 'vcf' or 'vcf.gz' suffix, and that
the column names correspond to typical VCF column headers.
}
\examples{
x <- system.file("extdata/umccrise/snv/somatic-ensemble-PASS.vcf.gz", package = "gpgr")
(y <- is_vcf(x))
tmp_file <- tempfile(pattern = "fakeFile", fileext = "vcf")
writeLines(c("col1\tcol2\tcol3", "1\t2\t3"), con = tmp_file)
(z <- is_vcf(tmp_file))
}
