% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linx.R
\name{linx_breakend_process}
\alias{linx_breakend_process}
\title{Process LINX Breakend File}
\usage{
linx_breakend_process(x)
}
\arguments{
\item{x}{Path to \code{linx.breakend.tsv} file.}
}
\value{
The input file as a tibble with column description.
}
\description{
Processes the \code{linx.breakend.tsv} file.
}
