% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linx.R
\name{linx_clusters_process}
\alias{linx_clusters_process}
\title{Process LINX Clusters File}
\usage{
linx_clusters_process(x)
}
\arguments{
\item{x}{Path to \code{linx.clusters.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Processes the \code{linx.clusters.tsv} file.
}
