% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linx.R
\name{linx_drivercatalog_read}
\alias{linx_drivercatalog_read}
\title{Read LINX Driver Catalog File}
\usage{
linx_drivercatalog_read(x)
}
\arguments{
\item{x}{Path to \code{linx.driver.catalog.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Reads the \code{linx.driver.catalog.tsv} file.
}
\examples{
x <- system.file("extdata/linx/tables/linx.driver.catalog.tsv.gz", package = "gpgr")
(l <- linx_drivercatalog_read(x))
}
