% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linx.R
\name{linx_drivers_process}
\alias{linx_drivers_process}
\title{Process LINX Drivers File}
\usage{
linx_drivers_process(x)
}
\arguments{
\item{x}{Path to \code{linx.drivers.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Processes the \code{linx.drivers.tsv} file.
}
