% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linx.R
\name{linx_drivers_read}
\alias{linx_drivers_read}
\title{Read LINX Drivers File}
\usage{
linx_drivers_read(x)
}
\arguments{
\item{x}{Path to \code{linx.drivers.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Reads the \code{linx.drivers.tsv} file.
}
\examples{
x <- system.file("extdata/linx/tables/linx.drivers.tsv", package = "gpgr")
(l <- linx_drivers_read(x))
}
