% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linx.R
\name{linx_fusion_process}
\alias{linx_fusion_process}
\title{Process LINX Fusion File}
\usage{
linx_fusion_process(x)
}
\arguments{
\item{x}{Path to \code{linx.fusion.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Processes the \code{linx.fusion.tsv} file.
}
