% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linx.R
\name{linx_links_process}
\alias{linx_links_process}
\title{Process LINX Links File}
\usage{
linx_links_process(x)
}
\arguments{
\item{x}{Path to \code{linx.links.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Processes the \code{linx.links.tsv} file.
}
\examples{
x <- system.file("extdata/linx/tables/linx.links.tsv.gz", package = "gpgr")
(l <- linx_links_process(x))
}
