% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linx.R
\name{linx_svs_process}
\alias{linx_svs_process}
\title{Process LINX SVS File}
\usage{
linx_svs_process(x)
}
\arguments{
\item{x}{Path to \code{linx.svs.tsv} file.}
}
\value{
The input file as a processed tibble.
}
\description{
Processes the \code{linx.svs.tsv} file.
}
