% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linx.R
\name{linx_vissegments_read}
\alias{linx_vissegments_read}
\title{Read LINX VisSegments File}
\usage{
linx_vissegments_read(x)
}
\arguments{
\item{x}{Path to \code{linx.vis_segments.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Reads the \code{linx.vis_segments.tsv} file.
}
\examples{
x <- system.file("extdata/linx/tables/linx.vis_segments.tsv", package = "gpgr")
(l <- linx_vissegments_read(x))
}
