% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linx_vis.R
\name{linx_vissvdata_read}
\alias{linx_vissvdata_read}
\title{Read LINX VisSvData File}
\usage{
linx_vissvdata_read(x)
}
\arguments{
\item{x}{Path to \code{linx.vis_sv_data.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Reads the \code{linx.vis_sv_data.tsv} file.
}
\examples{
x <- system.file("extdata/linx/tables/linx.vis_sv_data.tsv.gz", package = "gpgr")
(l <- linx_vissvdata_read(x))
}
