% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkdir}
\alias{mkdir}
\title{Make Directory}
\usage{
mkdir(d)
}
\arguments{
\item{d}{Directory to create.}
}
\value{
If directory exists, do nothing. If it doesn't, create it and return
invisibly a logical indicating if the operation succeeded.
}
\description{
Creates a directory.
}
