% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\name{plot_bnd_sr_pr_tot_lines}
\alias{plot_bnd_sr_pr_tot_lines}
\title{Line plot for SR, PR and SR + PR for BNDs}
\usage{
plot_bnd_sr_pr_tot_lines(
  d,
  title = "SR, PR and SR + PR line plot for BNDs",
  subtitle = "Events are sorted by decreasing tot values."
)
}
\arguments{
\item{d}{A data.frame with an SR_PR_alt column.}

\item{title}{Main title of plot.}

\item{subtitle}{Subtitle of plot.}
}
\value{
A ggplot2 plot object.
}
\description{
Plots the number of split reads (\code{SR}), paired end reads (\code{PR}), and their
sum (\code{tot}) across all BNDs, sorted by \code{tot}.
}
\examples{
x <- system.file("extdata/umccrise/sv/manta.tsv", package = "gpgr")
d <- process_sv(x)$unmelted
plot_bnd_sr_pr_tot_lines(d)
}
