% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{process_purple_cnv_germline}
\alias{process_purple_cnv_germline}
\title{Process PURPLE CNV germline File for UMCCRISE}
\usage{
process_purple_cnv_germline(x, v = "2.39")
}
\arguments{
\item{x}{Path to \code{purple.cnv.germline.tsv} file.}

\item{v}{PURPLE version (default: 2.39). Used to determine the column names.}
}
\value{
List with two elements:
\itemize{
\item \code{tab}: Tibble with more condensed columns.
\item \code{descr}: Description of tibble columns.
}
}
\description{
Processes the \code{purple.cnv.germline.tsv} file.
and selects columns of interest.
}
\examples{
x <- system.file("extdata/purple/v2.39/purple.cnv.germline.tsv", package = "gpgr")
(pp <- process_purple_cnv_germline(x)$tab)

}
