% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{purple_cnv_germ_process}
\alias{purple_cnv_germ_process}
\title{Process PURPLE CNV germline File for UMCCRISE}
\usage{
purple_cnv_germ_process(x)
}
\arguments{
\item{x}{Path to \code{purple.cnv.germline.tsv} file.}
}
\value{
List with two elements:
\itemize{
\item \code{tab}: Tibble with more condensed columns.
\item \code{descr}: Description of tibble columns.
}
}
\description{
Processes the \code{purple.cnv.germline.tsv} file.
and selects columns of interest.
}
\examples{
x <- system.file("extdata/purple/purple.cnv.germline.tsv", package = "gpgr")
(pp <- purple_cnv_germ_process(x))
}
