% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{purple_cnv_germline_read}
\alias{purple_cnv_germline_read}
\title{Read PURPLE CNV Germline File}
\usage{
purple_cnv_germline_read(x)
}
\arguments{
\item{x}{Path to \code{purple.cnv.germline.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Reads the \code{purple.cnv.germline.tsv} file.
}
\examples{
x <- system.file("extdata/purple/purple.cnv.germline.tsv", package = "gpgr")
(p <- purple_cnv_germline_read(x))

}
