% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{purple_cnv_som_gene_process}
\alias{purple_cnv_som_gene_process}
\title{Process PURPLE CNV Gene File for UMCCRISE}
\usage{
purple_cnv_som_gene_process(x, g = NULL)
}
\arguments{
\item{x}{Path to \code{purple.cnv.gene.tsv} file.}

\item{g}{Path to gene file containing at least three columns:
\itemize{
\item \code{symbol}: gene name (character).
\item \code{tumorsuppressor}: is this gene a tumor suppressor (TRUE/FALSE).
\item \code{oncogene}: is this gene an oncogene (TRUE/FALSE).
}}
}
\value{
List with two elements:
\itemize{
\item \code{tab}: Tibble filtered to genes found in  \code{g}.
\item \code{descr}: Description of tibble columns.
}
}
\description{
Processes the \code{purple.cnv.gene.tsv} file. Keeps genes that are in the
\href{https://github.com/umccr/genes/blob/893a655801ce92715f05517b5052e4e81904e870/panels/umccr_2019-03-20.tsv}{UMCCR cancer gene list}
and selects columns of interest.
}
\examples{
x <- system.file("extdata/purple/purple.cnv.gene.tsv", package = "gpgr")
g <- system.file("extdata/ref/umccr_cancer_genes_2019-03-20.tsv", package = "gpgr")
(pp <- purple_cnv_som_gene_process(x, g))

}
