% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{purple_cnv_som_gene_read}
\alias{purple_cnv_som_gene_read}
\title{Read PURPLE CNV Gene File}
\usage{
purple_cnv_som_gene_read(x)
}
\arguments{
\item{x}{Path to \code{purple.cnv.gene.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Reads the \code{purple.cnv.gene.tsv} file, which summarises copy number
alterations of each gene in the HMF panel
(see \href{https://github.com/hartwigmedical/hmftools/tree/master/purity-ploidy-estimator#gene-copy-number-file}{this table}).
}
\examples{
x <- system.file("extdata/purple/purple.cnv.gene.tsv", package = "gpgr")
(p <- purple_cnv_som_gene_read(x))

}
