% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{purple_cnv_som_read}
\alias{purple_cnv_som_read}
\title{Read PURPLE CNV Somatic File}
\usage{
purple_cnv_som_read(x)
}
\arguments{
\item{x}{Path to \code{purple.cnv.somatic.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Reads the \code{purple.cnv.somatic.tsv} file, which contains the copy number
profile of all (contiguous) segments of the tumor sample
(see \href{https://github.com/hartwigmedical/hmftools/tree/master/purple#copy-number-file}{this table}).
}
\examples{
x <- system.file("extdata/purple/purple.cnv.somatic.tsv", package = "gpgr")
(p <- purple_cnv_som_read(x))
}
