% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{purple_purity_read}
\alias{purple_purity_read}
\title{Read PURPLE Purity file}
\usage{
purple_purity_read(x)
}
\arguments{
\item{x}{Path to the \code{purple.purity.tsv} file.}
}
\value{
The input file as a tibble and a summarised tibble with a
description of each metric.
}
\description{
Reads the \code{purple.purity.tsv} file containing a summary of the purity fit.
}
\examples{
x <- system.file("extdata/purple/purple.purity.tsv", package = "gpgr")
(p <- purple_purity_read(x))
}
