% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{purple_snv_vcf_read}
\alias{purple_snv_vcf_read}
\title{Read PURPLE Somatic SNV VCF}
\usage{
purple_snv_vcf_read(x)
}
\arguments{
\item{x}{Path to the \code{purple.somatic.vcf.gz} file.}
}
\value{
A list with the input file as a tibble (with specific INFO fields)
and a tibble with a description of each INFO field.
}
\description{
Reads the \code{purple.somatic.vcf.gz} file.
}
\examples{
x <- system.file("extdata/purple/purple.somatic.vcf.gz", package = "gpgr")
(snv <- purple_snv_vcf_read(x))

}
