% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{read_purple_cnv_gene}
\alias{read_purple_cnv_gene}
\title{Read PURPLE CNV Gene File}
\usage{
read_purple_cnv_gene(x, v)
}
\arguments{
\item{x}{Path to \code{purple.cnv.gene.tsv} file.}

\item{v}{PURPLE version. Used to determine the column names.}
}
\value{
The input file in a tibble.
}
\description{
Reads the \code{purple.cnv.gene.tsv} file, which summarises copy number
alterations of each gene in the HMF panel
(see \href{https://github.com/hartwigmedical/hmftools/tree/master/purity-ploidy-estimator#gene-copy-number-file}{this table}).
}
\examples{
x <- system.file("extdata/purple/v2.39/purple.cnv.gene.tsv", package = "gpgr")
read_purple_cnv_gene(x, "2.39")

}
