% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{read_purple_cnv_germline}
\alias{read_purple_cnv_germline}
\title{Read PURPLE CNV Germline File}
\usage{
read_purple_cnv_germline(x)
}
\arguments{
\item{x}{Path to \code{purple.cnv.germline.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Reads the \code{purple.cnv.germline.tsv} file.
}
\examples{
x <- system.file("extdata/purple/purple.cnv.germline.tsv", package = "gpgr")
(p <- read_purple_cnv_germline(x))

}
