% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{read_purple_purity}
\alias{read_purple_purity}
\title{Read PURPLE Purity file}
\usage{
read_purple_purity(x)
}
\arguments{
\item{x}{Path to the \code{purple.purity.tsv} file.}
}
\value{
A named character vector containing a summary of the purity fit.
}
\description{
Reads the \code{purple.purity.tsv} file containing a summary of the purity fit.
}
\examples{
x <- system.file("extdata/purple/purple.purity.tsv", package = "gpgr")
(p <- read_purple_purity(x))

}
