% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{read_purple_qc}
\alias{read_purple_qc}
\title{Read PURPLE QC file}
\usage{
read_purple_qc(x)
}
\arguments{
\item{x}{Path to the \code{purple.qc} file.}
}
\value{
A named character vector containing QC values for several PURPLE
metrics.
}
\description{
Reads the \code{purple.qc} file containing.
}
\examples{
x <- system.file("extdata/purple/v2.39/purple.qc", package = "gpgr")
p <- read_purple_qc(x)
p

}
