% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{read_purple_version}
\alias{read_purple_version}
\title{Read PURPLE version file}
\usage{
read_purple_version(x)
}
\arguments{
\item{x}{Path to the \code{purple.version} file.}
}
\value{
A list with the elements:
\itemize{
\item \code{version}: The version of PURPLE used.
\item \code{build_date}: The build date of the PURPLE version.
}
}
\description{
Reads the \code{purple.version} file containing the PURPLE version and build date.
}
\examples{
x <- system.file("extdata/purple/purple.version", package = "gpgr")
(v <- read_purple_version(x))

}
