% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\name{read_sv_tsv}
\alias{read_sv_tsv}
\title{Read SV TSV}
\usage{
read_sv_tsv(x, v = "0.18")
}
\arguments{
\item{x}{Path to \code{manta.tsv} output by umccrise.}

\item{v}{umccrise version (default: 0.18). Used to verify the column names.}
}
\value{
A tibble
}
\description{
Reads the TSV file output by umccrise.
}
\examples{
x <- system.file("extdata/umccrise/v0.18/sv/manta.tsv", package = "gpgr")
(sv <- read_sv_tsv(x))

}
