% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord.R
\name{run_chord}
\alias{run_chord}
\title{Run CHORD}
\usage{
run_chord(
  vcf.snv = NULL,
  vcf.sv = NULL,
  df.sv = NULL,
  sample.name = NULL,
  ref.genome = "hg38",
  sv.caller = "manta",
  ...
)
}
\arguments{
\item{vcf.snv}{Path to VCF containing SNVs and INDELs.}

\item{vcf.sv}{Path to VCF containing SVs.}

\item{df.sv}{A data.frame object containing the columns
'SVTYPE' and 'SVLEN' from a Manta SV VCF.}

\item{sample.name}{Name of sample to use.}

\item{ref.genome}{Human genome assembly. One of 'hg38' (default), 'hg19' or 'GRCh37'.}

\item{sv.caller}{manta (default) or gridss.}

\item{...}{Other arguments to be passed to \code{CHORD::extractSigsChord()}.}
}
\value{
List with extracted signatures and HRD prediction.
}
\description{
Runs CHORD given SNV and SV VCF files. \strong{NOTE}: make sure you have
the BSgenome.Hsapiens.UCSC.hgXX installed.
}
\examples{
\dontrun{
snv <- system.file("extdata/umccrise/v0.18/snv/somatic-ensemble-PASS.vcf.gz", package = "gpgr")
sv <- system.file("extdata/umccrise/v0.18/sv/manta.vcf.gz", package = "gpgr")
chord_res <- run_chord(vcf.snv = snv, vcf.sv = sv, sample.name = "foo")
chord_res2 <- run_chord(vcf.snv = snv, df.sv = gpgr:::chord_mantavcf2df(sv), sample.name = "foo")
}


}
