% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.R
\name{session_info_kable}
\alias{session_info_kable}
\title{Session Information Kable}
\usage{
session_info_kable(pkgs)
}
\arguments{
\item{pkgs}{Vector of R packages to display in the vignette.}
}
\value{
A list with two kables containing information about the platform and
the specified packages (\code{pkgs}).
}
\description{
Session information kables for vignettes.
}
