% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_contribution_table}
\alias{sig_contribution_table}
\title{Create COSMIC Signature Table}
\usage{
sig_contribution_table(contr, type, outdir)
}
\arguments{
\item{contr}{A tibble with Rank, Signature, Contribution and RelFreq columns.}

\item{type}{One of Sig (old COSMIC), SBS, DBS or ID.}

\item{outdir}{Relative directory to write signature plots to for incorporating
in an RMarkdown report.}
}
\value{
The \code{contr} tibble with an additional \code{Plot} column pointing to
the local path to the corresponding signature plot (in markdown syntax).
}
\description{
Creates a COSMIC signature table for display in an RMarkdown report.
}
