% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_count_dbs}
\alias{sig_count_dbs}
\title{Count DBS Contexts}
\usage{
sig_count_dbs(vcf_gr)
}
\arguments{
\item{vcf_gr}{GRanges containing all mutation types from a single sample.}
}
\value{
A tibble containing the number of DBS per COSMIC context per gr.
}
\description{
Counts DBS Contexts.
}
