% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_count_indel}
\alias{sig_count_indel}
\title{Count INDEL Contexts}
\usage{
sig_count_indel(vcf_gr, ref_genome)
}
\arguments{
\item{vcf_gr}{GRanges containing all mutation types from a single sample.}

\item{ref_genome}{The BSGenome reference genome to use.}
}
\value{
A tibble containing the number of INDELs per COSMIC context per gr.
}
\description{
Counts INDEL Contexts.
}
