% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_count_snv}
\alias{sig_count_snv}
\title{Count SNV Contexts}
\usage{
sig_count_snv(vcf_gr, ref_genome)
}
\arguments{
\item{vcf_gr}{GRanges containing all mutation types from a single sample.}

\item{ref_genome}{The BSGenome reference genome to use.}
}
\value{
A list with two elements:
\itemize{
\item snv_counts: matrix containing the number of SNVs per COSMIC context per gr.
\item gr_snv: GRanges object containing the SNVs.
}
}
\description{
Counts SNV Contexts.
}
