% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_plot_dbs}
\alias{sig_plot_dbs}
\title{Plot DBS Mutation Characteristics}
\usage{
sig_plot_dbs(vcf_gr)
}
\arguments{
\item{vcf_gr}{GRanges containing all mutation types from a single sample.}
}
\value{
A list with two ggplot2 objects:
\itemize{
\item p_dbs_main: the main DBS contexts.
\item p_dbs_cont: the DBS contexts.
}
}
\description{
Plots DBS mutation characteristics.
}
