% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_plot_indel}
\alias{sig_plot_indel}
\title{Plot INDEL Mutation Characteristics}
\usage{
sig_plot_indel(vcf_gr, ref_genome)
}
\arguments{
\item{vcf_gr}{GRanges containing all mutation types from a single sample.}

\item{ref_genome}{The BSGenome reference genome object.}
}
\value{
A list with two ggplot2 objects:
\itemize{
\item p_indel_main: the main INDEL contexts.
\item p_indel_cont: the INDEL contexts.
}
}
\description{
Plots INDEL mutation characteristics.
}
