% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_workflow_run}
\alias{sig_workflow_run}
\title{Run MutationalPatterns Workflow}
\usage{
sig_workflow_run(vcf, sample_nm, ref_genome, outdir)
}
\arguments{
\item{vcf}{VCF file with SNVs.}

\item{sample_nm}{Sample name.}

\item{ref_genome}{The BSGenome reference genome to use.}

\item{outdir}{Directory path to output all the results to.}
}
\description{
Runs a MutationalPatterns Workflow.
}
\details{
Writes plots and signature contribution tables to \code{outdir}.
}
