% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tsv_is_empty}
\alias{tsv_is_empty}
\title{Does the TSV file contain any non-header rows?}
\usage{
tsv_is_empty(
  x,
  comment = "##",
  col_types = readr::cols(.default = "c"),
  n_max = 1,
  ...
)
}
\arguments{
\item{x}{Path to TSV file.}

\item{comment}{Taken from \code{\link[readr:read_delim]{readr::read_tsv()}}.}

\item{col_types}{Taken from \code{\link[readr:read_delim]{readr::read_tsv()}}.}

\item{n_max}{Taken from \code{\link[readr:read_delim]{readr::read_tsv()}}.}

\item{...}{Other arguments to be passed to \code{\link[readr:read_delim]{readr::read_tsv()}}.}
}
\value{
TRUE if there is at least one non-header row in the TSV, FALSE otherwise.
}
\description{
Checks if the TSV file contains any non-header rows.
}
\examples{

tmp1 <- tempfile()
writeLines(c("col1\tcol2\tcol3", "1\t2\t3"), con = tmp1)
(a <- tsv_is_empty(tmp1))

tmp2 <- tempfile()
writeLines(c("col1\tcol2\tcol3"), con = tmp2)
(b <- tsv_is_empty(tmp2))

tmp3 <- tempfile()
writeLines(c("##meta1", "##meta2", "col1\tcol2\tcol3"), con = tmp3)
(c <- tsv_is_empty(tmp3))

}
