% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcf_is_empty}
\alias{vcf_is_empty}
\title{Does the VCF file contain any variants?}
\usage{
vcf_is_empty(x)
}
\arguments{
\item{x}{Path to VCF file.}
}
\value{
TRUE if there is at least one variant in the VCF, FALSE otherwise.
}
\description{
Checks if the VCF file contains any variants.
}
\examples{
x <- system.file("extdata/umccrise/snv/somatic-ensemble-PASS.vcf.gz", package = "gpgr")
(y <- vcf_is_empty(x))

tmp1 <- tempfile(pattern = "fakeFile", fileext = "vcf")
writeLines(c("col1\tcol2\tcol3", "1\t2\t3"), con = tmp1)
\dontrun{
vcf_is_empty(tmp1)
}

vcf_cols <- c("#CHROM", "POS", "ID", "REF", "ALT", "QUAL",
              "FILTER", "INFO", "FORMAT")
tmp2 <- tempfile(pattern = "fakeFile", fileext = "vcf")
writeLines(paste(vcf_cols, collapse = "\t"), con = tmp2)
(z <- vcf_is_empty(tmp2))

tmp3 <- tempfile(pattern = "fakeFile", fileext = "FOO")
writeLines(paste(vcf_cols, collapse = "\t"), con = tmp3)
\dontrun{
vcf_is_empty(tmp3)
}

}
