% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoviral.R
\name{virusbreakend_summary_read}
\alias{virusbreakend_summary_read}
\title{Read VIRUSBreakend Summary File}
\usage{
virusbreakend_summary_read(x)
}
\arguments{
\item{x}{Path to \code{virusbreakend.vcf.summary.tsv} file.}
}
\value{
List with two elements:
\itemize{
\item \code{tab}: Tibble containing data.
\item \code{descr}: Description of tibble columns.
}
}
\description{
Reads the \code{virusbreakend.vcf.summary.tsv} file.
}
\examples{
x <- system.file("extdata/virusbreakend/virusbreakend.vcf.summary.tsv", package = "gpgr")
(vb <- virusbreakend_summary_read(x))
}
