% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_jsongz}
\alias{write_jsongz}
\title{Write gzipped JSON}
\usage{
write_jsongz(x, path, ...)
}
\arguments{
\item{x}{An object to be serialized to JSON.}

\item{path}{File on disk (should end in '.gz').}

\item{...}{Additional arguments passed to \code{\link[jsonlite:read_json]{jsonlite::write_json()}}}
}
\description{
Serializes an object to JSON and writes to a gzipped file.
}
