% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_tsvgz}
\alias{write_tsvgz}
\title{Write a data frame to a tab delimited gzipped file}
\usage{
write_tsvgz(x, file, ...)
}
\arguments{
\item{x}{A data frame or tibble to write to disk.}

\item{file}{File or connection to write to (should end in '.gz').}

\item{...}{Additional arguments passed to \code{\link[readr:write_delim]{readr::write_tsv()}}.}
}
\value{
Returns the input \code{x} invisibly.
}
\description{
Writes a data frame to a tab delimited gzipped file.
}
