% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_tsvjsongz}
\alias{write_tsvjsongz}
\title{Write a data frame to gzipped TSV and JSON files}
\usage{
write_tsvjsongz(x, path, maindir)
}
\arguments{
\item{x}{The data frame to write.}

\item{path}{Relative path to write the files to, sans the file extensions -
these will be appended appropriately.}

\item{maindir}{Main directory to write the files to.}
}
\description{
Writes a data frame to gzipped TSV and JSON files. Files
will be written to \verb{maindir/<path>.tsv.gz} and \verb{maindir/json/<path>.json.gz}.
}
