% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{cancer_rmd}
\alias{cancer_rmd}
\title{Generate UMCCR Cancer Report}
\usage{
cancer_rmd(
  af_global,
  af_keygenes,
  batch_name,
  bcftools_stats,
  conda_list,
  dragen_hrd,
  img_dir,
  key_genes,
  oncokb_genes,
  virusbreakend_tsv,
  virusbreakend_vcf,
  purple_purity,
  purple_qc,
  purple_som_cnv_ann,
  purple_som_cnv,
  purple_som_gene_cnv,
  purple_som_snv_vcf,
  somatic_snv_vcf,
  somatic_snv_summary,
  somatic_sv_tsv,
  somatic_sv_vcf,
  result_outdir,
  tumor_name,
  mutpat_dir = NULL,
  hrdetect_file = NULL,
  chord_file = NULL,
  out_file = NULL,
  quiet = FALSE
)
}
\arguments{
\item{af_global}{Path to \code{af_tumor.txt} file.}

\item{af_keygenes}{Path to \code{af_tumor_keygenes.txt} file.}

\item{batch_name}{Name of batch sample.}

\item{bcftools_stats}{Path to \code{bcftools_stats.txt} file.}

\item{conda_list}{Path to \code{conda_pkg_list.txt} file.}

\item{dragen_hrd}{Path to DRAGEN \code{hrdscore.csv} file.}

\item{img_dir}{Path to directory containing PURPLE plots.}

\item{key_genes}{Path to UMCCR cancer gene file.}

\item{oncokb_genes}{Path to OncoKB database file.}

\item{virusbreakend_tsv}{Path to VIRUSBreakend summary file.}

\item{virusbreakend_vcf}{Path to VIRUSBreakend VCF file.}

\item{purple_purity}{Path to \code{purple.purity.tsv}.}

\item{purple_qc}{Path to \code{purple.qc}.}

\item{purple_som_cnv_ann}{Path to annotated and prioritised \code{purple.cnv.somatic.tsv}.}

\item{purple_som_cnv}{Path to \code{purple.cnv.somatic.tsv}.}

\item{purple_som_gene_cnv}{Path to \code{purple.cnv.gene.tsv}.}

\item{purple_som_snv_vcf}{Path to \code{purple.somatic.vcf.gz}.}

\item{somatic_snv_vcf}{Path to \code{somatic-PASS.vcf.gz} SNV VCF.}

\item{somatic_snv_summary}{Path to \code{somatic_snv_summary.json} JSON.}

\item{somatic_sv_tsv}{Path to SV TSV file.}

\item{somatic_sv_vcf}{Path to SV VCF file.}

\item{result_outdir}{Path to directory to write tidy JSON/TSV results.}

\item{tumor_name}{Name of tumor sample.}

\item{mutpat_dir}{Path to mutational patterns directory.}

\item{hrdetect_file}{Path to HRDetect file.}

\item{chord_file}{Path to CHORD file.}

\item{out_file}{Path to output HTML file (needs '.html' suffix) (def: \verb{\{tumor_name\}_cancer_report.html}).}

\item{quiet}{Suppress log printing during rendering.}
}
\value{
Path to rendered HTML report.
}
\description{
Generates a UMCCR Cancer Report. It does so with the following steps:
\enumerate{
\item move the img_dir into 'tmp/img_dir'
\item copy the rmd into 'tmp/cancer_report.Rmd'
\item render the rmd inside 'tmp/'
\item return the path to the output HTML
}
}
