% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord.R
\name{chord_run}
\alias{chord_run}
\title{Run CHORD}
\usage{
chord_run(
  vcf.snv = NULL,
  vcf.sv = NULL,
  df.sv = NULL,
  sample.name = NULL,
  ref.genome = "hg38",
  sv.caller = "manta",
  outpath = NULL,
  ...
)
}
\arguments{
\item{vcf.snv}{Path to VCF containing SNVs and INDELs.}

\item{vcf.sv}{Path to VCF containing SVs.}

\item{df.sv}{A data.frame object containing the columns
'SVTYPE' and 'SVLEN' from a Manta SV VCF.}

\item{sample.name}{Name of sample to use.}

\item{ref.genome}{Human genome assembly. One of 'hg38' (default), 'hg19' or 'GRCh37'.}

\item{sv.caller}{manta (default) or gridss.}

\item{outpath}{File to write CHORD predictions to on disk
(should end in '.gz'). If not specified, results won't be written to disk.}

\item{...}{Other arguments to be passed to \code{\link[CHORD:extractSigsChord]{CHORD::extractSigsChord()}}.}
}
\value{
List with extracted signatures and HRD prediction.
}
\description{
Runs CHORD given SNV and SV VCF files. \strong{NOTE}: make sure you have
the BSgenome.Hsapiens.UCSC.hgXX installed.
}
\examples{

snv <- system.file("extdata/umccrise/snv/somatic-ensemble-PASS.vcf.gz", package = "gpgr")
sv <- system.file("extdata/umccrise/sv/manta.vcf.gz", package = "gpgr")
chord_res <- chord_run(
  vcf.snv = snv, df.sv = chord_mantavcf2df(sv),
  sample.name = "foo"
)
# chord_res2 <- chord_run(vcf.snv = snv, vcf.sv = sv, sample.name = "foo",
#                         outpath = "nogit/chord_results.json.gz")
}
