% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umccrise.R
\name{hrd_results_tabs}
\alias{hrd_results_tabs}
\title{HRDetect and CHORD Summary Table}
\usage{
hrd_results_tabs(hrdetect_res, chord_res, dragen_res)
}
\arguments{
\item{hrdetect_res}{Result from running \code{\link[sigrap:hrdetect_run]{sigrap::hrdetect_run()}}.}

\item{chord_res}{Result from running \code{\link[sigrap:chord_run]{sigrap::chord_run()}}.}

\item{dragen_res}{Result from running \code{\link[=dragen_hrd]{dragen_hrd()}}.}
}
\value{
A list with a tibble and a gt_tbl object (see \code{\link[gt:gt]{gt::gt()}}).
}
\description{
Return a summary table with HRDetect and CHORD results.
}
\examples{
\dontrun{
snv <- system.file("extdata/umccrise/snv/somatic-ensemble-PASS.vcf.gz", package = "gpgr")
sv <- system.file("extdata/umccrise/sv/manta.vcf.gz", package = "gpgr")
cnv <- system.file("extdata/purple/purple.cnv.somatic.tsv", package = "gpgr")
nm <- "SampleA"
genome <- "hg38"
snvoutdir <- tempdir()
hrdetect_res <- sigrap::hrdetect_run(nm, snv, sv, cnv, genome, snvoutdir)
chord_res <- sigrap::chord_run(
  vcf.snv = snv, sample.name = nm,
  df.sv = gpgr:::chord_mantavcf2df(sv)
)
dragen_res <- gpgr::dragen_hrd("path/to/sample.hrdscore.csv")
hrd_results_tabs(hrdetect_res = hrdetect_res, chord_res = chord_res, dragen_res = dragen_res)
}

}
