% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrdetect.R
\name{hrdetect_read_purple_cnv}
\alias{hrdetect_read_purple_cnv}
\title{Read PURPLE Somatic CNVs for HRDetect}
\usage{
hrdetect_read_purple_cnv(x)
}
\arguments{
\item{x}{Path to \code{purple.cnv.somatic.tsv} file.}
}
\value{
Tibble containing following columns:
\itemize{
\item chromosome, start, end
\item copyNumber (total)
\item minorAllelePloidy
}
}
\description{
Reads PURPLE somatic CNVs for use with HRDetect.
}
\examples{
x <- system.file("extdata/purple/purple.cnv.somatic.tsv", package = "gpgr")
(cnv <- hrdetect_read_purple_cnv(x))
}
