% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrdetect.R
\name{hrdetect_read_snvindel_vcf}
\alias{hrdetect_read_snvindel_vcf}
\title{Read VCF with SNVs/INDELs for use with HRDetect}
\usage{
hrdetect_read_snvindel_vcf(x)
}
\arguments{
\item{x}{Path to VCF.}
}
\value{
List containing CHROM, POS, REF and ALT columns
for SNVs and INDELs in separate tibbles.
}
\description{
Reads a VCF with SNVs/INDELs for use with HRDetect.
}
\examples{
x <- system.file("extdata/umccrise/snv/somatic-ensemble-PASS.vcf.gz", package = "gpgr")
(l <- hrdetect_read_snvindel_vcf(x))
}
