% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrdetect.R
\name{hrdetect_read_sv_vcf}
\alias{hrdetect_read_sv_vcf}
\title{Read VCF with SVs for use with HRDetect}
\usage{
hrdetect_read_sv_vcf(x, nm = NULL, genome = "hg38")
}
\arguments{
\item{x}{Path to VCF.}

\item{nm}{Sample name.}

\item{genome}{Human genome version (default: hg38. hg19 means GRCh37).}
}
\value{
Tibble with following BEDPE-like columns:
\itemize{
\item chrom1, start1, end1
\item chrom2, start2, end2
\item sample
\item strand1, strand2
}
}
\description{
Reads a VCF with SVs for use with HRDetect.
}
\examples{
x <- system.file("extdata/umccrise/sv/manta.vcf.gz", package = "gpgr")
sv_bedpe <- hrdetect_read_sv_vcf(x, nm = "SAMPLE")
head(sv_bedpe)
}
