% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrdetect.R
\name{hrdetect_run}
\alias{hrdetect_run}
\title{Run HRDetect via signature.tools.lib}
\usage{
hrdetect_run(
  nm,
  snvindel_vcf,
  sv_vcf,
  cnv_tsv,
  genome = "hg38",
  snvoutdir = tempdir(),
  sigsToUse = c(1, 2, 3, 5, 6, 8, 13, 17, 18, 20, 26, 30),
  outpath = NULL
)
}
\arguments{
\item{nm}{Sample name.}

\item{snvindel_vcf}{Path to VCF with SNVs and INDELs.}

\item{sv_vcf}{Path to VCF with SVs.}

\item{cnv_tsv}{Path to \code{purple.cnv.somatic.tsv} file.}

\item{genome}{Human genome version (default: hg38. hg19 means GRCh37).}

\item{snvoutdir}{Directory to output SNV signature analysis results.}

\item{sigsToUse}{COSMIC SNV signatures to use.}

\item{outpath}{File to write HRDetect predictions to on disk
(should end in '.gz'). If not specified, results won't be written to disk.}
}
\value{
Tibble with sample name and HRD probability in first two columns.
}
\description{
Runs HRDetect as described in the
\href{https://github.com/Nik-Zainal-Group/signature.tools.lib}{signature.tools.lib repository}.
}
\examples{
snvindel_vcf <- system.file(
  "extdata/umccrise/snv/somatic-ensemble-PASS.vcf.gz",
  package = "gpgr"
)
sv_vcf <- system.file("extdata/umccrise/sv/manta.vcf.gz", package = "gpgr")
cnv_tsv <- system.file("extdata/purple/purple.cnv.somatic.tsv", package = "gpgr")
nm <- "SampleA"
genome <- "hg38"
snvoutdir <- tempdir()
(res <- hrdetect_run(nm, snvindel_vcf, sv_vcf, cnv_tsv, genome, snvoutdir))
# hrdetect_run(nm, snvindel_vcf, sv_vcf, cnv_tsv, genome, snvoutdir,
#              outpath = "nogit/hrdetect_results.json.gz")
}
