% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linx.R
\name{linx_drivercatalog_process}
\alias{linx_drivercatalog_process}
\title{Process LINX Driver Catalog File}
\usage{
linx_drivercatalog_process(x)
}
\arguments{
\item{x}{Path to \code{linx.driver.catalog.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Processes the \code{linx.driver.catalog.tsv} file.
}
