% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{purple_cnv_som_gene_process}
\alias{purple_cnv_som_gene_process}
\title{Process PURPLE CNV Gene File for UMCCRISE}
\usage{
purple_cnv_som_gene_process(x, g = NULL)
}
\arguments{
\item{x}{Path to \code{purple.cnv.gene.tsv} file.}

\item{g}{Path to gene file containing at least three columns:
\itemize{
\item \code{ensembl_gene_symbol}: gene name (character).
\item \code{tsgene}: is this gene a tumor suppressor (TRUE/FALSE).
\item \code{oncogene}: is this gene an oncogene (TRUE/FALSE).
}}
}
\value{
List with two elements:
\itemize{
\item \code{tab}: Tibble filtered to genes found in  \code{g}.
\item \code{descr}: Description of tibble columns.
}
}
\description{
Processes the \code{purple.cnv.gene.tsv} file. Keeps genes that are in the
UMCCR cancer gene list
(\href{https://raw.githubusercontent.com/umccr/gene_panels/v24.03.0/somatic_panel/3_final_panel/final_panel.tsv}{v24.03.0}) and selects columns of interest.
}
\examples{
x <- system.file("extdata/purple/purple.cnv.gene.tsv", package = "gpgr")
g <- system.file("extdata/ref/somatic_panel-v24.03.0.tsv", package = "gpgr")
(pp <- purple_cnv_som_gene_process(x, g))
}
