% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{purple_qc_read}
\alias{purple_qc_read}
\title{Read PURPLE QC file}
\usage{
purple_qc_read(x)
}
\arguments{
\item{x}{Path to the \code{purple.qc} file.}
}
\value{
The input file as a tibble and a summarised tibble with a
description of each metric.
}
\description{
Reads the \code{purple.qc} file.
}
\examples{
x <- system.file("extdata/purple/purple.qc", package = "gpgr")
(q <- purple_qc_read(x))
}
