% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{read_purple_cnv_germline}
\alias{read_purple_cnv_germline}
\title{Read PURPLE CNV Germline File}
\usage{
read_purple_cnv_germline(x, v = "2.39")
}
\arguments{
\item{x}{Path to \code{purple.cnv.germline.tsv} file.}

\item{v}{PURPLE version (default: 2.39). Used to determine the column names.}
}
\value{
The input file as a tibble.
}
\description{
Reads the \code{purple.cnv.germline.tsv} file.
}
\examples{
x <- system.file("extdata/purple/v2.39/purple.cnv.germline.tsv", package = "gpgr")
(p <- read_purple_cnv_germline(x, "2.39"))

}
