% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purple.R
\name{read_purple_cnv_somatic}
\alias{read_purple_cnv_somatic}
\title{Read PURPLE CNV Somatic File}
\usage{
read_purple_cnv_somatic(x)
}
\arguments{
\item{x}{Path to \code{purple.cnv.somatic.tsv} file.}
}
\value{
The input file as a tibble.
}
\description{
Reads the \code{purple.cnv.somatic.tsv} file, which contains the copy number
profile of all (contiguous) segments of the tumor sample
(see \href{https://github.com/hartwigmedical/hmftools/tree/master/purity-ploidy-estimator#copy-number-file}{this table}).
}
\examples{
x <- system.file("extdata/purple/v2.39/purple.cnv.somatic.tsv", package = "gpgr")
(p <- read_purple_cnv_somatic(x))

}
