% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\name{read_sv_tsv}
\alias{read_sv_tsv}
\title{Read SV TSV}
\usage{
read_sv_tsv(x)
}
\arguments{
\item{x}{Path to \code{manta.tsv} output by umccrise.}
}
\value{
A tibble corresponding to the input TSV file.
}
\description{
Reads the Manta TSV file output by umccrise.
}
\examples{
x <- system.file("extdata/umccrise/sv/manta.tsv", package = "gpgr")
(sv <- read_sv_tsv(x))

}
