% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord.R
\name{run_chord}
\alias{run_chord}
\title{Run CHORD}
\usage{
run_chord(snv, sv, sample, genome = "hg38")
}
\arguments{
\item{snv}{Path to SNV VCF.}

\item{sv}{Path to SV VCF.}

\item{sample}{Name of sample to use.}

\item{genome}{Human genome assembly. One of 'hg38' (default) or 'hg19'.}
}
\value{
List with extracted signatures and HRD prediction.
}
\description{
Runs CHORD given SNV and SV VCF files. \strong{NOTE}: make sure you have
the BSgenome.Hsapiens.UCSC.hgXX installed.
}
\examples{
\dontrun{
snv <- system.file("extdata/umccrise/v0.18/snv/somatic-ensemble-PASS.vcf.gz", package = "gpgr")
sv <- system.file("extdata/umccrise/v0.18/sv/manta.vcf.gz", package = "gpgr")
chord_res <- run_chord(snv = snv, sv = sv, sample = "foo")
}


}
