% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord.R
\name{run_chord}
\alias{run_chord}
\title{Run CHORD}
\usage{
run_chord(snv, sv, sample, genome = "hg38")
}
\arguments{
\item{snv}{Path to SNV VCF.}

\item{sv}{Path to SV VCF.}

\item{sample}{Name of sample to use.}

\item{genome}{Human genome assembly. One of 'hg38' (default) or 'hg19'.}
}
\value{
List with extracted signatures and HRD prediction.
}
\description{
Runs CHORD given SNV and SV VCF files. \strong{NOTE}: you need to
load the following R packages:
BSgenome.Hsapiens.UCSC.hg38, mutSigExtractor, and CHORD.
}
