% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationalpatterns.R
\name{sig_plot_snv}
\alias{sig_plot_snv}
\title{Plot SNV Mutation Characteristics}
\usage{
sig_plot_snv(gr_snv, snv_counts, ref_genome)
}
\arguments{
\item{gr_snv}{GRanges containing SNVs from a single sample.}

\item{snv_counts}{A matrix with counts of SNV contexts.}

\item{ref_genome}{The BSGenome reference genome to use.}
}
\value{
A list with four ggplot2 objects:
\itemize{
\item p_heatmap: a SNV mutation matrix as a heatmap.
This is especially usefull when looking at a wide mutational context.
\item p_river: a SNV mutation matrix as a riverplot.
This is especially usefull when looking at a wide mutational context.
\item p_96_profile: relative contribution of 96 trinucleotides.
\item p_spectrum: point mutation spectrum.
}
}
\description{
Plots SNV mutation characteristics.
}
