% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\name{split_double_col}
\alias{split_double_col}
\title{Split Two-Field Column}
\usage{
split_double_col(.data, col, is_pct = FALSE)
}
\arguments{
\item{.data}{Input tibble data.}

\item{col}{Column to split.}

\item{is_pct}{Multiply by 100 (logical).}
}
\value{
A modified tibble with two columns.
}
\description{
Splits a column with 2 comma-separated fields into
two columns.
}
\examples{
x <- tibble::tibble(a = letters[1:10], b = paste0(round(runif(10), 2), ",", round(runif(10), 2)))
(s <- gpgr:::split_double_col(x, b))

}
